refactor(ws-client): dual node/browser entrypoints, migrate to vitest
