+++ packages/ws-client/package.json
@@ -21,8 +21,14 @@
   "type": "module",
   "exports": {
     ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./dist/index.js"
+      "node": {
+        "types": "./dist/node.d.ts",
+        "default": "./dist/node.js"
+      },
+      "default": {
+        "types": "./dist/browser.d.ts",
+        "default": "./dist/browser.js"
+      }
     }
   },
   "engines": {
@@ -36,10 +42,11 @@
     "@types/ws": "^8.5.4",
     "get-port": "^6.1.2",
     "http-terminator": "^3.2.0",
-    "jest": "^30.0.0"
+    "undici": "^6.0.0",
+    "vitest": "^4.0.16"
   },
   "scripts": {
-    "test": "NODE_OPTIONS=--experimental-vm-modules jest",
+    "test": "vitest run",
     "build": "tsgo --build tsconfig.build.json"
   }
 }
