+++ packages/ws-client/src/browser.ts
@@ -1,0 +1,2 @@
+export { WebSocketKeepAlive } from './keepalive-stub.js'
+export { CloseCode, DisconnectError } from './keepalive-shared.js'
