+++ packages/ws-client/src/keepalive-shared.ts
@@ -1,0 +1,15 @@
+export class DisconnectError extends Error {
+  constructor(
+    public wsCode: CloseCode = CloseCode.Policy,
+    public xrpcCode?: string,
+  ) {
+    super()
+  }
+}
+
+// https://www.rfc-editor.org/rfc/rfc6455#section-7.4.1
+export enum CloseCode {
+  Normal = 1000,
+  Abnormal = 1006,
+  Policy = 1008,
+}
