+++ packages/ws-client/src/node.ts
@@ -1,0 +1,2 @@
+export { WebSocketKeepAlive } from './keepalive.js'
+export { CloseCode, DisconnectError } from './keepalive-shared.js'
