+++ packages/ws-client/tests/keepalive.test.ts
@@ -3,9 +3,10 @@ import { createServer } from 'node:http'
 import type { AddressInfo } from 'node:net'
 // eslint-disable-next-line import/default
 import httpTerminator from 'http-terminator'
+import { describe, expect, it } from 'vitest'
 import { WebSocketServer } from 'ws'
 import { wait } from '@atproto/common'
-import { CloseCode, WebSocketKeepAlive } from '../src/index.js'
+import { CloseCode, WebSocketKeepAlive } from '../src/node.ts'
 
 describe('WebSocketKeepAlive', () => {
   it('uses a heartbeat to reconnect if a connection is dropped', async () => {
