+++ packages/ws-client/tsconfig.build.json
@@ -5,5 +5,6 @@
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./dist",
+    "lib": ["ES2023", "DOM", "DOM.Iterable"],
   },
 }
