+++ packages/ws-client/tsconfig.test.json
@@ -1,8 +1,8 @@
 {
-  "extends": ["../../tsconfig/jest.tsconfig.json"],
+  "extends": ["../../tsconfig/vitest.tsconfig.json"],
   "include": ["./tests", "./src/**/*.test.ts"],
-  "references": [{ "path": "./tsconfig.build.json" }],
   "compilerOptions": {
     "rootDir": ".",
   },
+  "references": [{ "path": "./tsconfig.build.json" }],
 }
